currentPersistenceIds
