+++ core/src/test/scala/akka/persistence/r2dbc/query/CurrentPersistenceIdsQuerySpec.scala
@@ -2,7 +2,7 @@
  * Copyright (C) 2021 Lightbend Inc. <https://www.lightbend.com>
  */
 
-package akka.persistence.r2dbc.state
+package akka.persistence.r2dbc.query
 
 import scala.concurrent.duration._
 
@@ -10,19 +10,19 @@ import akka.Done
 import akka.actor.testkit.typed.scaladsl.LogCapturing
 import akka.actor.testkit.typed.scaladsl.ScalaTestWithActorTestKit
 import akka.actor.typed.ActorSystem
+import akka.persistence.query.PersistenceQuery
 import akka.persistence.r2dbc.TestActors
-import akka.persistence.r2dbc.TestActors.DurableStatePersister
+import akka.persistence.r2dbc.TestActors.Persister
 import akka.persistence.r2dbc.TestConfig
 import akka.persistence.r2dbc.TestData
 import akka.persistence.r2dbc.TestDbLifecycle
-import akka.persistence.r2dbc.state.scaladsl.R2dbcDurableStateStore
-import akka.persistence.state.DurableStateStoreRegistry
+import akka.persistence.r2dbc.query.scaladsl.R2dbcReadJournal
 import akka.persistence.typed.PersistenceId
 import akka.stream.scaladsl.Sink
 import com.typesafe.config.ConfigFactory
 import org.scalatest.wordspec.AnyWordSpecLike
 
-class PersistenceIdsQuerySpec
+class CurrentPersistenceIdsQuerySpec
     extends ScalaTestWithActorTestKit(
       ConfigFactory
         .parseString("""
@@ -36,8 +36,7 @@ class PersistenceIdsQuerySpec
 
   override def typedSystem: ActorSystem[_] = system
 
-  private val store = DurableStateStoreRegistry(testKit.system)
-    .durableStateStoreFor[R2dbcDurableStateStore[String]](R2dbcDurableStateStore.Identifier)
+  private val query = PersistenceQuery(testKit.system).readJournalFor[R2dbcReadJournal](R2dbcReadJournal.Identifier)
 
   private val zeros = "0000"
   private val entityTypeHint = nextEntityTypeHint()
@@ -50,22 +49,23 @@ class PersistenceIdsQuerySpec
 
     val probe = createTestProbe[Done]
     pids.foreach { pid =>
-      val persister = spawn(TestActors.DurableStatePersister(pid))
-      persister ! DurableStatePersister.PersistWithAck("s-1", probe.ref)
-      persister ! DurableStatePersister.Stop(probe.ref)
+      val persister = spawn(TestActors.Persister(pid))
+      persister ! Persister.PersistWithAck("e-1", probe.ref)
+      persister ! Persister.PersistWithAck("e-2", probe.ref)
+      persister ! Persister.Stop(probe.ref)
     }
 
-    probe.receiveMessages(numberOfPids * 2, 30.seconds) // ack + stop done
+    probe.receiveMessages(numberOfPids * 3, 30.seconds) // 2 acks + stop done
   }
 
-  "Durable State persistenceIds" should {
+  "Event Sourced currentPersistenceIds" should {
     "retrieve all ids" in {
-      val result = store.persistenceIds().runWith(Sink.seq).futureValue
+      val result = query.currentPersistenceIds().runWith(Sink.seq).futureValue
       result shouldBe pids.map(_.id)
     }
 
     "retrieve ids afterId" in {
-      val result = store.currentPersistenceIds(afterId = Some(pids(9).id), limit = 7).runWith(Sink.seq).futureValue
+      val result = query.currentPersistenceIds(afterId = Some(pids(9).id), limit = 7).runWith(Sink.seq).futureValue
       result shouldBe pids.slice(10, 17).map(_.id)
     }
